possibly fixed dupe bug
